# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API Minha Folga
  version: 1.1.0
  summary: Cadastro de interesse, preferências, atendimento, ferramentas da Bia (Hal-AI) e administração.
  description: |
    Contrato versionado entre `frontend/` (Vue) e `backend/` (Node/Express). Todas as rotas vivem sob `/api`,
    na mesma origem do site. Regras gerais:

    - Erros sempre em JSON: `{ "error": { "code", "message", "fields"? , ...extra }, "requestId" }`.
      `message` é texto em português que pode ser exibido ao usuário; `code` é estável.
    - Chamadas do navegador com método inseguro exigem `Content-Type: application/json` e, se houver
      cabeçalho `Origin`, que ela esteja na lista permitida (defesa CSRF de base).
    - Corpos com campos inesperados são recusados (`validation_error`). Limite de corpo: 16 KB em todas as rotas.
    - Web Push usa autorização independente de navegador, endpoint cifrado e token técnico do dono; ver notifications.openapi.yaml.
    - Sessões usam cookies `HttpOnly; Secure; SameSite=Strict` definidos pelo servidor. Nenhum token vai para
      `localStorage`, query string ou pixels. Dados pessoais nunca aparecem em URL.
    - Respostas de cadastro e de pedido de saída são genéricas: não revelam se um número ou CPF já está cadastrado.
    - Em `PRE_LAUNCH`, toda rota `/api/credit/*` responde 403 `credit_phase_locked`.

    Decisão D1 (docs/DECISOES.md, 26/09/2026): este sistema **não mexe com WhatsApp**. Número oficial, Bia, webchat,
...
servers:
...
paths:
...
  /api/admin/auth/login:
    post:
      tags: [ admin ]
      summary: Primeiro fator (e-mail e senha). Abre sessão parcial que exige MFA.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false, required: [ email, password ], properties: { email: { type: string, format: email, maxLength: 200 }, password: { type: string, maxLength: 200 } } }
      responses:
        '200':
          description: 'Senha correta; próximo passo: `mfa_required` ou `mfa_setup_required`. Cookie mf_admin (Path=/api/admin).'
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminLoginStep' } } }
        '401': { description: '`invalid_credentials` (genérico, inclusive conta bloqueada/desativada).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/admin/auth/mfa/setup:
    post:
      tags: [ admin ]
      summary: Inicia o cadastro do TOTP (somente sessão parcial de usuário sem MFA). Exige X-CSRF-Token.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        '200': { description: Segredo e QR para o aplicativo autenticador., content: { application/json: { schema: { $ref: '#/components/schemas/MfaSetup' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: '`mfa_already_enabled`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/admin/auth/mfa/verify:
    post:
      tags: [ admin ]
      summary: Segundo fator (código TOTP). Conclui o cadastro do MFA quando pendente e libera a sessão.
      description: |
        Falhas de MFA têm contador próprio por conta (5 a cada 15 minutos), que um novo login com senha correta não
        zera. A quinta falha bloqueia a conta por 15 minutos (429 `rate_limited`); conta bloqueada não confere código.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false, required: [ code ], properties: { code: { type: string, pattern: '^\d{6}$' } } }
      responses:
        '200': { description: Sessão completa., content: { application/json: { schema: { $ref: '#/components/schemas/AdminSession' } } } }
        '400': { description: '`invalid_code`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/admin/auth/me:
    get:
      tags: [ admin ]
      summary: Sessão atual (usuário, permissões e token CSRF). Sessão parcial devolve o passo pendente.
      responses:
        '200': { description: Sessão., content: { application/json: { schema: { oneOf: [ { $ref: '#/components/schemas/AdminSession' }, { $ref: '#/components/schemas/AdminLoginStep' } ] } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/admin/auth/logout:
    post:
      tags: [ admin ]
      summary: Encerra a sessão administrativa.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        '204': { description: Encerrada. }
  /api/admin/auth/password:
    post:
      tags: [ admin ]
      summary: Troca a própria senha (mínimo 12 caracteres). Obrigatória no primeiro acesso.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false, required: [ currentPassword, newPassword ], properties: { currentPassword: { type: string, maxLength: 200 }, newPassword: { type: string, minLength: 12, maxLength: 200 } } }
      responses:
        '204': { description: Senha alterada; outras sessões do usuário são encerradas. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/admin/leads:
    get:
      tags: [ admin ]
      summary: Lista cadastros com filtros (permissão leads:read). Telefone, CPF e e-mail sempre mascarados.
      description: |
        Nome, empregador, faixas e UF/cidade visíveis para leads:read. Cada consulta é auditada (`admin.leads_listed`, com
        filtros, página e quantidade; sem dados do titular).
      parameters:
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { in: query, name: state, schema: { $ref: '#/components/schemas/LeadState' } }
        - { in: query, name: source, schema: { type: string } }
        - { in: query, name: employmentType, schema: { $ref: '#/components/schemas/EmploymentType' } }
        - { in: query, name: jobTenure, schema: { $ref: '#/components/schemas/JobTenure' } }
        - { in: query, name: incomeRange, schema: { $ref: '#/components/schemas/IncomeRange' } }
        - { in: query, name: uf, schema: { $ref: '#/components/schemas/Uf' } }
        - { in: query, name: phoneVerified, schema: { type: string, enum: [ 'true', 'false' ] }, description: 'Telefone validado (a própria pessoa falou com a Bia).' }
        - { in: query, name: purpose, schema: { type: string, enum: [ launch_notice, marketing ] } }
        - { in: query, name: granted, schema: { type: string, enum: [ 'true', 'false' ] } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200':
          description: Página de cadastros.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AdminLead' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/leads/{id}:
    get:
      tags: [ admin ]
      summary: Detalhe do cadastro com trilha de eventos e submissões pendentes de revisão, mascaradas (leads:read).
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Detalhe., content: { application/json: { schema: { $ref: '#/components/schemas/AdminLeadDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/leads/{id}/reveal-contact:
    post:
      tags: [ admin ]
      summary: Revela telefone, e-mail e CPF (leads:read_contact). Exige justificativa; registrado em auditoria (sem os valores).
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false, required: [ reason ], properties: { reason: { type: string, minLength: 5, maxLength: 200 } } }
      responses:
        '200': { description: Contato e documento., content: { application/json: { schema: { $ref: '#/components/schemas/RevealedContact' } } } }
        '409': { description: '`lead_anonymized`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/admin/leads/export:
    post:
      tags: [ admin ]
      summary: Exporta CSV filtrado (leads:export). Justificativa obrigatória; auditado. Telefone, CPF e e-mail mascarados.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ reason ]
              properties:
                reason: { type: string, minLength: 5, maxLength: 200 }
                filters: { type: object, additionalProperties: { type: string } }
      responses:
        '200': { description: CSV (text/csv; charset=utf-8). }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/admin/privacy-requests:
    get:
      tags: [ admin ]
      summary: Pedidos de titulares (privacy:manage).
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [ open, in_progress, fulfilled, rejected ] } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
      responses:
        '200':
          description: Lista.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PrivacyRequest' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [ admin ]
      summary: Registra pedido recebido por outro canal (privacy:manage).
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ requestType, channel, summary, receivedAt ]
              properties:
                requestType: { $ref: '#/components/schemas/PrivacyRequestType' }
                leadId: { type: [ string, 'null' ], format: uuid }
                supportRequestId: { type: [ string, 'null' ], format: uuid }
                channel: { type: string, maxLength: 60 }
                summary: { type: string, minLength: 5, maxLength: 1000 }
                receivedAt: { type: string, format: date-time }
      responses:
        '201': { description: Registrado., content: { application/json: { schema: { $ref: '#/components/schemas/PrivacyRequest' } } } }
  /api/admin/privacy-requests/{id}:
    patch:
      tags: [ admin ]
      summary: Atualiza/cumpre o pedido. Ações irreversíveis exigem confirmação e são auditadas.
      description: |
        `action`: `revoke_all` (revoga todas as finalidades), `anonymize_lead` (remove nome, telefone, CPF, e-mail, empregador
        e cidade do cadastro, apaga as submissões guardadas para revisão com o mesmo telefone ou CPF, substitui dados de
        todas as solicitações de atendimento do titular — vinculadas ao cadastro, a do pedido e as com a mesma chave de
        contato, como as abertas pela Bia ou pelo site com o mesmo WhatsApp ou e-mail —, encerra sessões e cancela
        mensagens pendentes; os ids alcançados ficam na auditoria) e `suppress_contact` (impede novos envios ao
        número e do CPF, pelas chaves originais, mesmo após anonimização). Preferências do titular prevalecem.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                status: { type: string, enum: [ open, in_progress, fulfilled, rejected ] }
                resolution: { type: string, maxLength: 1000 }
                action: { type: string, enum: [ revoke_all, anonymize_lead, suppress_contact ] }
                confirm: { type: boolean, description: 'Obrigatório (true) para anonymize_lead.' }
      responses:
        '200': { description: Atualizado., content: { application/json: { schema: { $ref: '#/components/schemas/PrivacyRequest' } } } }
        '400': { description: '`confirmation_required` ou validação.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    adminSession: { type: apiKey, in: cookie, name: mf_admin }
    contactSession: { type: apiKey, in: cookie, name: mf_contact }
    agentApiKey: { type: apiKey, in: header, name: X-API-Key }
  parameters:
    Csrf:
      in: header
      name: X-CSRF-Token
      required: true
      description: Token da sessão administrativa (de /api/admin/auth/me ou do login). Obrigatório em métodos inseguros do admin.
      schema: { type: string }
  responses:
    ValidationError:
      description: '`validation_error` com `fields` { campo: mensagem }.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: '`rate_limited`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: '`unauthorized` / `session_required`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: '`forbidden` (sem permissão; tentativa auditada).'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: '`not_found`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [ error ]
      properties:
        error:
          type: object
          required: [ code, message ]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            attemptsRemaining: { type: integer }
            retryAfterSeconds: { type: integer }
        requestId: { type: string }
...
    EmploymentType: { type: string, enum: [ clt, domestico, rural, outro ] }
    JobTenure: { type: string, enum: [ menos_3_meses, 3_a_12_meses, 1_a_3_anos, mais_3_anos ] }
    IncomeRange: { type: string, enum: [ ate_2000, 2000_a_4000, 4000_a_7000, acima_7000, nao_informar ] }
    Uf: { type: string, enum: [ AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN, RS, RO, RR, SC, SP, SE, TO ] }
...
    LeadState:
      type: string
      description: 'D1: received (site, telefone não validado) → verified (a Bia atestou o número) → invited; laterais unsubscribed e expired. verification_pending, waiting e human_support só em cadastros anteriores à D1.'
      enum: [ received, verified, invited, unsubscribed, expired, verification_pending, waiting, human_support ]
...
    AdminRole: { type: string, enum: [ admin, support, privacy, marketing ] }
    PrivacyRequestType: { type: string, enum: [ access, correction, deletion, revocation, opposition, portability, information, other ] }
...
    AdminLoginStep:
      type: object
      properties:
        status: { type: string, enum: [ mfa_required, mfa_setup_required ] }
        csrfToken: { type: string }
    MfaSetup:
      type: object
      properties:
        secret: { type: string, description: Base32, para digitação manual. }
        otpauthUrl: { type: string }
        qrSvg: { type: string, description: SVG gerado no servidor. }
    AdminSession:
      type: object
      properties:
        status: { const: authenticated }
        user: { $ref: '#/components/schemas/AdminUser' }
        permissions: { type: array, items: { type: string, enum: [ leads:read, leads:read_contact, leads:export, support:read, support:update, privacy:manage, metrics:read, audit:read, users:manage, team:read, collection:manage, notifications:send, notifications:manage ] } }
        csrfToken: { type: string }
    AdminUserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/AdminRole' }
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/AdminRole' }
        mfaEnabled: { type: boolean }
        mustChangePassword: { type: boolean }
        disabled: { type: boolean }
        lastLoginAt: { type: [ string, 'null' ], format: date-time }
        createdAt: { type: string, format: date-time }
    AdminLead:
      type: object
      properties:
        id: { type: string, format: uuid }
        preferredName: { type: string }
        fullName: { type: string }
        phoneHint: { type: string }
        phoneVerified: { type: boolean, description: 'Telefone validado pela Bia (número atestado).' }
        cpfHint: { type: [ string, 'null' ] }
        emailHint: { type: [ string, 'null' ] }
        employerName: { type: [ string, 'null' ] }
        employmentType: { $ref: '#/components/schemas/EmploymentType' }
        jobTenure: { type: [ string, 'null' ] }
        incomeRange: { type: [ string, 'null' ] }
        city: { type: [ string, 'null' ] }
        uf: { type: [ string, 'null' ] }
        interestTopic: { type: [ string, 'null' ] }
        state: { $ref: '#/components/schemas/LeadState' }
        source: { type: string }
        utmSource: { type: [ string, 'null' ] }
        utmCampaign: { type: [ string, 'null' ] }
        contactVerifiedAt: { type: [ string, 'null' ], format: date-time }
        createdAt: { type: string, format: date-time }
        purposes:
          type: object
          properties:
            launch_notice: { type: boolean }
            marketing: { type: boolean }
        pendingReview: { type: integer, description: 'Submissões guardadas para revisão (não aplicadas ao cadastro).' }
    AdminLeadDetail:
      allOf:
        - { $ref: '#/components/schemas/AdminLead' }
        - type: object
          properties:
            anonymized: { type: boolean }
            events:
              type: array
              items:
                type: object
                properties:
                  type: { type: string }
                  fromState: { type: [ string, 'null' ] }
                  toState: { type: [ string, 'null' ] }
                  purpose: { type: [ string, 'null' ] }
                  granted: { type: [ boolean, 'null' ] }
                  source: { type: string }
                  actorType: { type: string }
                  at: { type: string, format: date-time }
            pendingSubmissions:
              type: array
              description: 'Submissões do site ou da Bia que não sobrescreveram o cadastro (mascaradas).'
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  channel: { type: string, enum: [ site, bia ] }
                  source: { type: string }
                  reason: { type: string, enum: [ existing_phone, existing_cpf, existing_phone_and_cpf, cpf_conflict, cpf_mismatch, suppressed ] }
                  createdAt: { type: string, format: date-time }
                  data:
                    type: [ object, 'null' ]
                    properties:
                      fullName: { type: [ string, 'null' ] }
                      phoneHint: { type: [ string, 'null' ] }
                      cpfHint: { type: [ string, 'null' ] }
                      emailHint: { type: [ string, 'null' ] }
                      employerName: { type: [ string, 'null' ] }
                      employmentType: { type: [ string, 'null' ] }
                      jobTenure: { type: [ string, 'null' ] }
                      incomeRange: { type: [ string, 'null' ] }
                      city: { type: [ string, 'null' ] }
                      uf: { type: [ string, 'null' ] }
                      interestTopic: { type: [ string, 'null' ] }
                  consents:
                    type: object
                    properties:
                      launch_notice: { type: [ boolean, 'null' ] }
                      marketing: { type: [ boolean, 'null' ] }
                  applicable: { type: boolean, description: '`false` = só descartar (`cpf_conflict`, `suppressed` ou conteúdo ilegível).' }
...
    RevealedContact:
      type: object
      required: [ phone, email, cpf ]
      properties:
        phone: { type: string, description: E.164. }
        email: { type: [ string, 'null' ] }
        cpf: { type: [ string, 'null' ], description: '000.000.000-00; null em cadastros anteriores à D1.' }
...
    PrivacyRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        requestType: { $ref: '#/components/schemas/PrivacyRequestType' }
        leadId: { type: [ string, 'null' ] }
        supportRequestId: { type: [ string, 'null' ] }
        channel: { type: string }
        status: { type: string, enum: [ open, in_progress, fulfilled, rejected ] }
        summary: { type: string }
        resolution: { type: [ string, 'null' ] }
        receivedAt: { type: string, format: date-time }
        fulfilledAt: { type: [ string, 'null' ], format: date-time }
        handledBy: { oneOf: [ { $ref: '#/components/schemas/AdminUserRef' }, { type: 'null' } ] }
...